[WPE] Implement non-composited page rendering https://bugs.webkit.org/show_bug.cgi?id=303170
Reviewed by Carlos Garcia Campos.

This change implements the non-composited page rendering for WPE.

With this change, it's possible to utilize a simplified frame rendering
that does not use compositor at all. It's useful especially on low-end
embedded devices where GPU is not powerful or not present at all.

The non-composited mode implemented in this change can be enabled using
a -AcceleratedCompositing preference.

Additionally, this change adds a layout ref-test that protects the new
implementation as well as existing, GTK's implementation from
regressions.

Canonical link: https://commits.webkit.org/304678@main