Add Swift IPC layout test https://bugs.webkit.org/show_bug.cgi?id=303989 rdar://166293273
Reviewed by Richard Robinson.

We previously landed an ability for CoreIPC to call directly into Swift message
handlers, and we added a test message handler within the IPCTester
infrastructure. However, we landed no automated test to exercise this new code.
That's what this PR does.

This will only have any effect in builds where ENABLE_IPC_TESTING is true, which
as standard is ASAN or debug builds.

In all such builds, we'll now exercise a code path to call into C++ and receive
an asynchronous reply. If additionally ENABLE_IPC_TESTING_SWIFT is enabled,
we'll also call into a Swift message receiver.

As well as adding the new layout test (HTML) it also makes a tiny change to the
recently-added IPCTester logic, to ensure the C++ logic behaves slightly
differently to the Swift logic, as an extra layer of proof that the Swift
receiver was called. It also adds a missing call to removeMessageReceiver.

Test: ipc/swift-receiver.html
Canonical link: https://commits.webkit.org/304693@main