AX: iOS should let native focus handle presenting date time picker https://bugs.webkit.org/show_bug.cgi?id=304364 rdar://165526147
Reviewed by Tyler Wilcock and Chris Fleizach.

In 301528@main, date and datetime inputs started being treated as action elements. This
is fine for platforms where WebKit handles the creation and interaction of the date time
picker. But, on iOS, `showPicker` doesn't do anything. It instead relies on the peripheral
input system, which manages the presentation of the picker.

We should use the presence of segments to determine how we should handle the press, as exposed
through hasCustomFocusLogic on the date input. This will be different between the platforms,
as iOS doesn't have segments and doesn't manage it's own focus.

* Source/WebCore/accessibility/AccessibilityNodeObject.cpp:
(WebCore::isDateFieldWithStandardFocus):
(WebCore::nodeActionElement):
* Source/WebCore/html/HTMLInputElement.h:

Canonical link: https://commits.webkit.org/304700@main