Prevent Safari crashes when scrolling reddit https://bugs.webkit.org/show_bug.cgi?id=304296 rdar://166587473
Reviewed by Aditya Keerthi.

A crash on safari was reported and the stack traces point to the
FindTextMatchCallbackAggregator. The aggregator holds a WeakPtr
to a WebPageProxy, which can be null at the aggregator's destruction
(which is when the page is used).

To fix this issue, I store the WebPageProxy in a RefPtr and check
for valid state before using the page.

* Source/WebKit/UIProcess/FindTextMatchesCallbackAggregator.cpp:
(WebKit::FindTextMatchCallbackAggregator::~FindTextMatchCallbackAggregator):

Canonical link: https://commits.webkit.org/304702@main