Use ProcessID instead of pid_t in AccessibilityRemoteToken constructor https://bugs.webkit.org/show_bug.cgi?id=304491 rdar://problem/166875779
Reviewed by Joshua Hoffman.

The AccessibilityRemoteToken constructor was using pid_t directly, which is
a POSIX type not available on Windows. This caused a build failure on Windows.

Changed the constructor parameter type from pid_t to ProcessID, which is a
cross-platform type defined in WTF that uses int on Windows and pid_t on Unix.
This matches the member variable declaration which already used ProcessID.

* Source/WebCore/accessibility/AXObjectCache.h:
(WebCore::AccessibilityRemoteToken::AccessibilityRemoteToken):

Canonical link: https://commits.webkit.org/304754@main