Reorganize code in FormListedElement::formOwnerRemovedFromTree to get rid of suppressions https://bugs.webkit.org/show_bug.cgi?id=304476
Reviewed by Anne van Kesteren.

The suppressions were needed because clearing of m_formAttributeTargetObserver, which deletes
FormAttributeTargetObserver, made the for loop non-trivial.

Re-organize the code so that clearing of m_formAttributeTargetObserver happens in an outer scope.

Also wrap the code which checks if any of the ancestor nodes matches the current form element in
a lambda so that we don't need to suppress a warning for rootNode.

No new tests since there should be no behavioral differences.

* Source/WebCore/html/FormListedElement.cpp:
(WebCore::FormListedElement::formOwnerRemovedFromTree):

Canonical link: https://commits.webkit.org/304764@main