AX: Force accessibility/text-stitching-inside-links.html to perform text stitching within the link, fixing it after 304507@main broke it https://bugs.webkit.org/show_bug.cgi?id=304354 rdar://166728489
Reviewed by Joshua Hoffman.

In https://commits.webkit.org/304507@main, we made links atomic accessibility elements in more scenarios, but that broke
the ability to test stitching within links as accessibility/text-stitching-inside-links.html relied on. Add a button
to the link in this test to ensure it is exposed in the accessibility as its descendants, allowing us to exercise the
text stitching logic.

* LayoutTests/accessibility/text-stitching-inside-links-expected.txt:
* LayoutTests/accessibility/text-stitching-inside-links.html:
* LayoutTests/platform/ios/accessibility/text-stitching-inside-links-expected.txt:

Canonical link: https://commits.webkit.org/304784@main