Teed byte streams can be GCed even though their parent stream can still fulfil read requests rdar://166562238 https://bugs.webkit.org/show_bug.cgi?id=304253
Reviewed by Chris Dumez.

We have an issue that leads to GC of streams in case of byteStreamTee.
If the stream gets GCed, the readable stream reader will also be GCed and the read promises will not get fulfilled.
A reader of a teed stream should not be GCed if the source stream is active.
To fix that issue, whenever a source stream is visited by GC, we are marking its teed streams as reachable from opaque roots if the source stream is readable.
We make sure to use a lock for the weak pointers and nullify the weak pointer before the corresponding stream gets destroyed.

Covered by reenabled tests.

Canonical link: https://commits.webkit.org/304858@main