Assert that m_ptr is non-null in RefPtr's `operator->` https://bugs.webkit.org/show_bug.cgi?id=304490
Reviewed by Alan Baradlay.

Assert that m_ptr is non-null in RefPtr's `operator->`, to make it more
likely to catch bugs in debug builds. Unfortunately, we are not able to
use a RELEASE_ASSERT() due to performance constraints. However, a null
dereference is undefined behavior so it will not necessarily crash and get
caught unless we have an assertion in place.

* Source/WTF/wtf/RefPtr.h:
* Tools/WebKitTestRunner/InjectedBundle/InjectedBundle.cpp:
(WTR::InjectedBundle::textDidChangeInTextField):
(WTR::InjectedBundle::textFieldDidBeginEditing):
(WTR::InjectedBundle::textFieldDidEndEditing):

Canonical link: https://commits.webkit.org/304890@main