REGRESSION(288788@main): SVG doesn't render gradient fill when using paint-order https://bugs.webkit.org/show_bug.cgi?id=304582 rdar://166997630
Reviewed by Nikolas Zimmermann.

If the paint-order is "stroke fill" and the transparency layer compositing is
used the gradient will fill the stroke itself not inside it.

The fix is to use the transparency layer compositing only if the paint-order is
"fill stroke". Otherwise clip the context to the stroke then fill the whole
rectangle with the gradient.

* LayoutTests/svg/text/text-paint-order-stroke-fill-gradient-expected.svg: Added.
* LayoutTests/svg/text/text-paint-order-stroke-fill-gradient.svg: Added.
* Source/WebCore/rendering/svg/legacy/LegacyRenderSVGResourceGradient.cpp:
(WebCore::LegacyRenderSVGResourceGradient::applyResource):

Canonical link: https://commits.webkit.org/304903@main