Split AnimationValue and friends out of GraphicsLayer https://bugs.webkit.org/show_bug.cgi?id=304677
Reviewed by Sam Weinig.

Address a couple of existing FIXMEs asking for these classes to be
moved out of GraphicsLayer to enable us to reduce unsafeness in a
subsequent change (so we can clearly attribute blame in case something
goes amiss).

We did take the liberty of making GraphicLayerAnimationValue's
m_timingFunction const as part of this change as that's trivial enough.

Canonical link: https://commits.webkit.org/304936@main